sample-app/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0
prometheus-client==0.21.1
